Fix git fetching new tags
Added a bunch of flags to the fetch and checkout commands. First, we need to fetch the tags (`--tags`) and we need to force it (`-f`) in case there is a conflicting tag. I also forced the git checkout to the tag incase files have been changed in the repo preventing the `git checkout` to work.